[Core] Split out ContentTypeDomainMapper to de couple contentTypeService a bit
Also to be able to reuse this fucntionality in other helpers.
